requirements.txt
Click==7.0
cycler==0.10.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==3.0.1
numpy==1.15.3
Pillow==5.3.0
pyparsing==2.3.0
python-dateutil==2.7.5
six==1.11.0
Werkzeug==0.14.1
